+++ src/environments/environment.ts
@@ -1,5 +1,5 @@
 export const environment = {
   production: false,
-  version: '1.0.18',
+  version: '1.0.19',
   curseForgeApiKey: ''
 };